requirements.txt
django==5.2.8
pymysql==1.1.2
djangorestframework
drf-spectacular
django-filter

django-extensions
ipython
python-dotenv==1.1.1
django-import-export==4.3.14
openpyxl==3.1.5
celery==5.5.3
redis==7.0.1
pymilvus==2.4.6
marshmallow==3.21.3
requests==2.32.3
cryptography
pydantic
openai==1.55.3
httpx[socks]
markdown==3.7
xhtml2pdf==0.2.15
llama-index
llama-index-llms-deepseek
